separated About the Data page UI from agencies container
