+++ src/js/containers/router/RouterRoutes.js
@@ -30,7 +30,7 @@ const RecipientContainer = React.lazy(() => import('containers/recipient/Recipie
 const AgencyProfileV2 = React.lazy(() => import('containers/agency/v2/AgencyContainerV2').then((comp) => comp));
 const Covid19Container = React.lazy(() => import('containers/covid19/Covid19Container').then((comp) => comp));
 const DataSourcesAndMethodologiesPage = React.lazy(() => import('components/covid19/DataSourcesAndMethodologiesPage').then((comp) => comp));
-const AgenciesContainer = React.lazy(() => import('containers/aboutTheData/AgenciesContainer').then((comp) => comp));
+const AboutTheDataPage = React.lazy(() => import('components/aboutTheData/AboutTheDataPage').then((comp) => comp));
 const AgenciesDetailContainer = React.lazy(() => import('containers/aboutTheData/AgenciesDetailContainer').then((comp) => comp));
 const ErrorPage = React.lazy(() => import('components/errorPage/ErrorPage').then((comp) => comp));
 
@@ -170,7 +170,7 @@ export const routes = [
     },
     {
         path: '/about-the-data/agencies',
-        component: AgenciesContainer,
+        component: AboutTheDataPage,
         exact: true,
         hide: !kGlobalConstants.DEV
     },
